Updated classes to track newest version of SECOAUTH. This update closes issues #3, #4, #8, and #36 (infinite redirects). This revision changes the authorization and token endpoints to be /openidconnect/auth and /openidconnect/token, respectively.

+++ openid-connect-server/src/main/java/org/mitre/oauth2/service/impl/DefaultOAuth2ClientDetailsEntityService.java
@@ -82,8 +82,8 @@ public class DefaultOAuth2ClientDetailsEntityService implements ClientDetailsEnt
     public ClientDetailsEntity createClient(String clientId, String clientSecret, 
     		Set<String> scope, Set<String> grantTypes, String redirectUri, Set<GrantedAuthority> authorities,
     		Set<String> resourceIds,
-    		String name, String description, boolean allowRefresh, Long accessTokenTimeout, 
-    		Long refreshTokenTimeout, String owner) {
+    		String name, String description, boolean allowRefresh, Integer accessTokenTimeout, 
+    		Integer refreshTokenTimeout, String owner) {
 		
 		// TODO: check "owner" locally?
